backend/requirements.txt
asgiref==3.8.1
Django==4.2
sqlparse==0.5.3
typing_extensions==4.12.2
mongoengine==0.27.0
djangorestframework==3.14.0
django-rest-framework-mongoengine==3.4.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
weaviate-client==3.25.2
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
celery==5.3.4
redis==5.0.1
scenedetect==0.6.2
